feat: mutate the data in db on actions, additional checks
Now, the data in the db is mutated for images which are simply moved or
moved to trash. Added additional check for name collisions when moving
files. Instead of seeing shutil's errors, user will be prompted with a
new name for the image files which seems to cause trouble when being
moved.