feat: lists with menus abstractions, UX improvements
Created a ListWithMenu base class to house all the logic associated with
handling events and incapsulating functionality as well as providing a
neat API. More errors, which are being raised in case parameters for the
MainWindow don't conform to the requirements, planning to add similar
checks for the cli arguments as well. Added a "show in explorer" option
to open a directory from FoldersList in the system file explorer.